Update go-swagger godep

+++ Godeps/_workspace/src/github.com/go-swagger/go-swagger/httpkit/middleware/parameter.go
@@ -1,3 +1,17 @@
+// Copyright 2015 go-swagger maintainers
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package middleware
 
 import (
@@ -61,9 +75,9 @@ func (p *untypedParamBinder) typeForSchema(tpe, format string, items *spec.Items
 	case "integer":
 		switch format {
 		case "int8":
-			return reflect.TypeOf(int32(0))
+			return reflect.TypeOf(int8(0))
 		case "int16":
-			return reflect.TypeOf(int64(0))
+			return reflect.TypeOf(int16(0))
 		case "int32":
 			return reflect.TypeOf(int32(0))
 		case "int64":
